Moved Schema to treeModel namespace. Moved Command and AttributeCommand to command namespace.
