+++ tests/attributecommand.js
@@ -6,7 +6,7 @@
 'use strict';
 
 import Editor from '/ckeditor5/core/editor.js';
-import AttributeCommand from '/ckeditor5/core/attributecommand.js';
+import AttributeCommand from '/ckeditor5/core/command/attributecommand.js';
 import Text from '/ckeditor5/core/treemodel/text.js';
 import Range from '/ckeditor5/core/treemodel/range.js';
 import Position from '/ckeditor5/core/treemodel/position.js';
